added used components to the about box
